perfomance improvements - was hitting open files limit
open/close the indexer only once (per response)

commit only on close (in case there are errors/failure the whole date
range will be pulled again, as start date/last successful should not be
updated in such cases)

There are cases where the searcher (IndexReader) can keep older files
open; seems that we don't need the searcher anyways. Instead there's now
an id field build from a hash of the query that the search would use and
there's updateDocument instead of addDocument. Update document is using
the hash to identify the documents to be replaced.