redo handle stats
it's not correct to sum daily uniques to get monthly uniques (similarly
for names and totals).
If there are 2 visits to /handle/123/456 and 2 visits to
/handle/123/456?showFull on a concrete day; the number of visits on that
day is any number between 2-4.

The repo pdf report shows visits by country (pulled with period=month)
and total visits; obtained by summing* the per day totals. These did
differ a lot for 11372/LRT-1974 (in Dec 2021 report).

*though this does not gurantee correct result, it usually is pretty
close as visits are roughly 30min sessions.